Updated express and mongoose, restructured route files
