+++ server/package.json
@@ -16,11 +16,11 @@
     "cookie-parser": "~1.4.4",
     "csv-parse": "^4.8.8",
     "debug": "~2.6.9",
-    "express": "~4.16.1",
+    "express": "^4.16.4",
     "express-paginate": "^1.0.0",
     "http-errors": "~1.6.3",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.9.15",
+    "mongoose": "^5.9.18",
     "mongoose-findorcreate": "^3.0.0",
     "morgan": "~1.9.1",
     "passport": "^0.4.1",
