+++ server/src/auth/util.js
@@ -1,19 +1,6 @@
-import jwt from 'jsonwebtoken';
-
 import User from '../models/user';
 
-import { JWT_SECRET } from '../../config.json';
-
-export const generateToken = (req, res) => {
-  const {
-    user: { _id: id, admin },
-  } = req;
-  const token = jwt.sign({ id, admin }, JWT_SECRET, {
-    expiresIn: '24h',
-  });
-  res.json({ token });
-};
-
+// eslint-disable-next-line import/prefer-default-export
 export const profileHandler = (
   accessToken,
   refreshToken,
