+++ server/src/routes/api/auth.js
@@ -1,7 +1,7 @@
 import express from 'express';
 
-import { generateToken } from '../auth/jwt';
-import passport from '../auth/passport';
+import { generateToken } from '../../auth/jwt';
+import passport from '../../auth/passport';
 
 const router = express.Router();
