+++ server/src/routes/api/index.js
@@ -2,14 +2,15 @@ import express from 'express';
 import passport from 'passport';
 import swaggerUi from 'swagger-ui-express';
 
+import authRouter from './auth';
 import dataRouter from './data';
 
 import Flight, {
   deleteFlight,
   getFlightById,
   updateFlight,
-} from '../models/flight';
-import apiSpec from '../../openapi.json';
+} from '../../models/flight';
+import apiSpec from '../../../openapi.json';
 
 const router = express.Router();
 
@@ -17,9 +18,23 @@ router.get('/', (req, res) => {
   res.json({ message: 'API home page' });
 });
 router.use('/docs', swaggerUi.serve, swaggerUi.setup(apiSpec));
+router.use('/auth', authRouter);
 router.use('/data', dataRouter);
 
 router.use(passport.authenticate('jwt', { session: false }));
+
+router.get('/flights/:id', async (req, res, next) => {
+  const { id } = req.params;
+  try {
+    const flight = await getFlightById(id);
+    if (!flight) {
+      next('fdsfsfsdf');
+    }
+    res.json(flight);
+  } catch (err) {
+    res.sendStatus(500);
+  }
+});
 
 router.post('/flights', async (req, res) => {
   const user = req.user._id;
@@ -34,19 +49,6 @@ router.post('/flights', async (req, res) => {
     if (err.name === 'ValidationError') {
       res.sendStatus(400);
     }
-    res.sendStatus(500);
-  }
-});
-
-router.get('/flights/:id', async (req, res, next) => {
-  const { id } = req.params;
-  try {
-    const flight = await getFlightById(id);
-    if (!flight) {
-      next('fdsfsfsdf');
-    }
-    res.json(flight);
-  } catch (err) {
     res.sendStatus(500);
   }
 });
@@ -79,6 +81,7 @@ router.delete('/flights/:id', async (req, res, next) => {
 });
 
 router.use('*', (req, res) => {
+  console.log({ req });
   res.sendStatus(404);
 });
